0.1.0 - Full Rewrite
A full rewrite of the plugin using 1.21 as a base for all the new features like brigadier for example.
Nothing of this rewrite was tested yet, so there's a high chance it is just unusable at this point.
I will commit some new changes tomorrow, containing all the other important stuff, like game management, configuration, etc.